// backend/src/utils/types.ts
export type ApiResponse<T> = {
	success: boolean;
	message: string;
	data?: T;
	errors?: Record<string, unknown>;
};
// ...
export type BankType = z.infer<typeof BankSchema> & { id?: string };